refactor zip install flow a little
