+++ src-tauri/Cargo.toml
@@ -24,6 +24,7 @@ futures-locks = "0.7.0"
 tokio = "1.21.2"
 tempdir = "0.3.7"
 walkdir = "2.3.2"
+bytes = "1.2.1"
 
 [features]
 # by default Tauri runs in production mode
