+++ src-tauri/src/main.rs
@@ -5,7 +5,7 @@
 
 use checks::{check_if_folder_looks_like_pocket, start_connection_thread};
 use futures_locks::RwLock;
-use install_core::start_zip_thread;
+use install_zip::start_zip_thread;
 use serde::{Deserialize, Serialize};
 use std::collections::HashSet;
 use std::fs::{self};
@@ -16,7 +16,7 @@ use tauri::api::dialog;
 use tauri::{App, Window};
 use walkdir::{DirEntry, WalkDir};
 mod checks;
-mod install_core;
+mod install_zip;
 
 struct PocketSyncState(RwLock<PathBuf>);
 
@@ -61,7 +61,7 @@ async fn read_text_file(
 ) -> Result<String, ()> {
     let pocket_path = state.0.read().await;
     let path = pocket_path.join(path);
-    println!("reading text file: {:?}", &path);
+    // println!("reading text file: {:?}", &path);
     let video_json = fs::read_to_string(path).unwrap();
     Ok(video_json)
 }
